templataize Grid class, move functions definitions from stroke.hpp to stroke.cpp
