+++ src/image_processing.cpp
@@ -1,7 +1,5 @@
 // local
 #include "image_processing.hpp"
-#include "grid.hpp"
-#include "stroke.hpp"
 
 // libs
 // spdlog
@@ -44,45 +42,6 @@ cv::Mat CropImageToSymbol(const cv::Mat &input_mat) {
     bounding_box = bounding_box | cv::boundingRect(contours[i]);
   }
   return input_mat(bounding_box);
-}
-
-Grid PlaceOnGrid(std::vector<mathboard::Stroke> &strokes) {
-  // calculate boundaries of grid
-  cv::Point2f bot_right_corner{0, 0};
-  cv::Point2f top_left_corner{INFINITY, INFINITY};
-  for (std::size_t i = 0; i < strokes.size(); i++) {
-    const cv::Point2f line_pos = strokes[i].GetPosition();
-    const cv::Rect line_BB = strokes[i].GetBoundingBox();
-
-    if (bot_right_corner.x < line_pos.x + line_BB.width) {
-      bot_right_corner.x = line_pos.x + line_BB.width;
-    }
-    if (top_left_corner.x > line_pos.x) {
-      top_left_corner.x = line_pos.x;
-    }
-    if (bot_right_corner.y < line_pos.y + line_BB.height) {
-      bot_right_corner.y = line_pos.y + line_BB.height;
-    }
-    if (top_left_corner.y > line_pos.y) {
-      top_left_corner.y = line_pos.y;
-    }
-  }
-
-  // calculate avrage size of one image
-  cv::Size2f avrage_size_of_image{0.0f, 0.0f};
-  for (std::size_t i = 0; i < strokes.size(); i++) {
-    avrage_size_of_image.width += strokes[i].GetBoundingBox().size().width;
-    avrage_size_of_image.height += strokes[i].GetBoundingBox().size().height;
-  }
-  avrage_size_of_image.width /= static_cast<float>(strokes.size());
-  avrage_size_of_image.height /= static_cast<float>(strokes.size());
-
-  // create grid
-  Grid grid(top_left_corner, bot_right_corner, avrage_size_of_image);
-  for (std::size_t i = 0; i < strokes.size(); i++) {
-    grid.InsertStroke(&strokes[i]);
-  }
-  return grid;
 }
 
 } // namespace mathboard
